refactor(jobs): merge corporation and character killmails
in order to reduce data duplication and match with ESI design
this commit is removing both corporation_killmails and
character_killmails which were storing only killmail_id
and killmail_hash collected from entities.

killmail jobs has been updated to seed this new table instead.

finally, we rollback to a job is equal to an action pattern.
as a result, job responsible to collect killmail details is
non longer run per entity and don't use any token.

every time a killmail will be inserted, a job will be dispatched
to collect related data (trigger might be replaced by an observer).
